Adds Bundled UI to Repository
This commit adds the bundled UI to the
project. We do this so that it can be easily
added to the app Docker image when the image is
built. Prior to this commit, one has to
create the bundle using `make package-ui`.
However, that would only work if one has
Node.js dependencies installed, which may
not have been the case. This makes it easier
to get going with the Docker image and
to deploy applications and modify them
in production.